Run erc by launching eeschematic and controlling the gui
This is heavily inspired by prior work from the Productize, Scott Bezek
and the splitflap contributors.
Specifically, this file:
https://github.com/productize/kicad-automation-scripts/blob/master/src/eeschema/schematic.py

I wrote my own version in rust to avoid having to use python2 and having
to deal with pip or equivalent to pull in a number of python
dependencies.

This is just a beginning.

TODO:
* Better error handling
  - In particular, informative error messages for
    - Missing eeschema executable (install kicad)
  - Gui not responding as we expected
* More configurability, in particular for the timeouts
* Run drc too