Workaround intermitent error
This fixes a scenario I encounter regularly when running ninja on the
ferris repo.
It is hard to reproduce as there needs to be a heavy load but I have
found a way to reproduce it reliably:

In a window, run: `stress --cpu 16` (or however many cores your machine
has).
In 9 other windows, run in parallel:
```
repeat 10 {docker run -v $HOME/Documents/code/ferris:/workdir -t kicad_cli bash -c "kicad_cli run-drc 0.1/base/ferris.kicad_pcb --headless"}
```

This will always fail in at least one of the parallel windows before
this.

The reason is that the process being spawned is not proof that the
X window was actually lauched.

This "fix" is hacky but reliably fix the issue: however overloaded the
system is, giving it one second from the process being spawned kind of
guarantees that the X window can launch, so we can later reliably find
it with xdotool.